Set fixed width to push rod separation & right side isn't cut out
I thought I did have the width as a constant instead of variable!
Apparently not.. Width is set to how far apart the push rods are, so
40mm. Length is variable, height is (somewhat) fixed - the floor is
fixed, the empty space is the width of timing belt plus extra space.

The right side is now a complete wall. Originally I had it only fill
up below the bottom belt clamp[1], but gaping wall looks like a bug,
so it's all filled in now.

[1] It is designed that way so the timing belt snuggly fits between two
walls

CODE CLEANUP TO FOLLOW
It's monkey patched as it stands, esp. the "temp" module